+++ Process.h
@@ -17,8 +17,8 @@ public:
 	void setCore(int coreID);
 	bool isFinished() const;
 	std::string getCreationTime() const;
-	float getMemorySize() const;
-	float getPageSize() const;
+	int getMemorySize() const;
+	int getPageSize() const;
 
 	void execute(); // Executes one instruction
 
@@ -33,7 +33,7 @@ private:
 	int currentInstruction;
 	int totalInstructions;
 	
-	float memorySize;
-	float pageSize;
+	int memorySize;
+	int pageSize;
 };
